# specs/openapi/api.v1.yaml
openapi: "3.1.0"
info:
  title: "LongoX Platform API"
  version: "1.0.0"
  description: "LongoX API — low-code automation and internal tools platform"
  contact:
    name: "LongoX Support"
    url: "https://longox.io/support"
  license:
    name: "Proprietary"
servers:
  - url: "https://api.longox.io"
    description: "Production"
  - url: "https://staging-api.longox.io"
    description: "Staging"
  - url: "http://localhost:3000"
    description: "Development"
paths:
...
  /api/v1/connectors/{id}:
    get:
      operationId: getConnector
      tags: ["Connectors"]
      summary: "Get connector details"
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "Connector"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connector"
...
  /api/v1/connectors/{id}/install:
    post:
      operationId: installConnector
      tags: ["Connectors"]
      summary: "Install a connector"
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "Installed connector"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectorInstallation"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT Bearer token obtained from /auth/login or /auth/register"
...
  schemas:
...
    Connector:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
        name:
          type: string
        displayName:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        icon:
          type: string
          nullable: true
        category:
          type: string
        trustLevel:
          type: string
        scope:
          type: string
          nullable: true
        authType:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ConnectorInstallation:
      type: object
      properties:
        id:
          type: string
        tenantId:
          type: string
        connectorVersionId:
          type: string
        status:
          type: string
        installedAt:
          type: string
          format: date-time
        installedBy:
          type: string
          nullable: true
        config:
          type: object
        createdAt:
          type: string
          format: date-time
...